separated line data and partition function databases. Updated one of the tests for new value of constants. Renamed database.line_data_protocol to database.protocols as there is  more than one protocol in that file.
