+++ archnemesis/database/__init__.py
@@ -5,5 +5,3 @@ from __future__ import annotations #  for 3.9 compatability
 from . import datatypes
 from . import filetypes
 
-from .hitran import HITRAN
-from .line_database_protocol import LineDatabaseProtocol
